packages/client/pyproject.toml
[project]
name = "vllm-grpc-client"
version = "0.1.0"
requires-python = ">=3.12"
description = "Async gRPC client for the vllm-grpc frontend"
dependencies = [
    "grpcio>=1.65",
    "vllm-grpc-gen",
]

[tool.uv.sources]
vllm-grpc-gen = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vllm_grpc_client"]
